+++ src/main/java/net/silentchaos512/gear/api/GearApi.java
@@ -114,6 +114,10 @@ public final class GearApi {
         return MaterialManager.getValues(includeChildren);
     }
 
+    public static boolean isMaterial(ItemStack stack) {
+        return MaterialManager.from(stack) != null;
+    }
+
     //endregion
 
     //region Parts