+++ src/main/java/net/silentchaos512/gear/block/INamedContainerExtraData.java
@@ -1,0 +1,8 @@
+package net.silentchaos512.gear.block;
+
+import net.minecraft.inventory.container.INamedContainerProvider;
+import net.minecraft.network.PacketBuffer;
+
+public interface INamedContainerExtraData extends INamedContainerProvider {
+    void encodeExtraData(PacketBuffer buffer);
+}