+++ src/main/java/net/silentchaos512/gear/init/Registration.java
@@ -1,6 +1,7 @@
 package net.silentchaos512.gear.init;
 
 import net.minecraft.block.Block;
+import net.minecraft.enchantment.Enchantment;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.merchant.villager.VillagerProfession;
 import net.minecraft.inventory.container.ContainerType;
@@ -25,6 +26,7 @@ import java.util.stream.Collectors;
 public final class Registration {
     public static final DeferredRegister<Block> BLOCKS = create(ForgeRegistries.BLOCKS);
     public static final DeferredRegister<ContainerType<?>> CONTAINERS = create(ForgeRegistries.CONTAINERS);
+    public static final DeferredRegister<Enchantment> ENCHANTMENTS = create(ForgeRegistries.ENCHANTMENTS);
     public static final DeferredRegister<EntityType<?>> ENTITIES = create(ForgeRegistries.ENTITIES);
     public static final DeferredRegister<Item> ITEMS = create(ForgeRegistries.ITEMS);
     public static final DeferredRegister<GlobalLootModifierSerializer<?>> LOOT_MODIFIERS = create(ForgeRegistries.LOOT_MODIFIER_SERIALIZERS);
@@ -39,6 +41,7 @@ public final class Registration {
         IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
         BLOCKS.register(modEventBus);
         CONTAINERS.register(modEventBus);
+        ENCHANTMENTS.register(modEventBus);
         ENTITIES.register(modEventBus);
         ITEMS.register(modEventBus);
         LOOT_MODIFIERS.register(modEventBus);
@@ -49,6 +52,7 @@ public final class Registration {
 
         ModBlocks.register();
         ModContainers.register();
+        GearEnchantments.register();
         ModEntities.register();
         ModItems.register();
         ModLootStuff.init();